requirements.txt
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.5.2
pkg-resources==0.0.0
pydub==0.23.1
python-dotenv==0.10.1
six==1.12.0
Werkzeug==0.15.2
yarl==1.3.0
